# model: Add support for GLM 4.5 family of models (#14921)