--- a/convert_hf_to_gguf.py
+++ b/convert_hf_to_gguf.py
@@ -6610,7 +6610,7 @@ class Glm4MoeModel(TextModel):
         special_vocab._set_special_token("eom", tokenizer.get_added_vocab()["<|observation|>"])  # 151338
 
         # Patch broken chat template
-        if special_vocab.chat_template and "visible_text(m.content).endswith" in special_vocab.chat_template:
+        if isinstance(special_vocab.chat_template, str) and "visible_text(m.content).endswith" in special_vocab.chat_template:
             special_vocab.chat_template = special_vocab.chat_template.replace(
                 """{{ visible_text(m.content) }}\n{{- '/nothink' if (enable_thinking is defined and not enable_thinking and not visible_text(m.content).endswith("/nothink")) else '' -}}""",
                 """{% set content = visible_text(m.content) %}{{ content }}\n{{- '/nothink' if (enable_thinking is defined and not enable_thinking and not content.endswith("/nothink")) else '' -}}""")