--- a/src/llama-hparams.h
+++ b/src/llama-hparams.h
@@ -72,7 +72,8 @@ struct llama_hparams {
     float    expert_weights_scale = 0.0;
     bool     expert_weights_norm  = false;
     uint32_t expert_gating_func   = LLAMA_EXPERT_GATING_FUNC_TYPE_NONE;
-    uint32_t moe_every_n_layers   = 0;
+    uint32_t moe_every_n_layers       = 0;
+    uint32_t num_nextn_predict_layers = 0;
 
     float f_norm_eps;
     float f_norm_rms_eps;