Add support for the newly released GLM 4.5 family of models.

- https://huggingface.co/zai-org/GLM-4.5
- https://huggingface.co/zai-org/GLM-4.5-Air
- https://z.ai/blog/glm-4.5

## Core Architecture

- Architecture Registration: Added LLM_ARCH_GLM45 enum and architecture mappings
- Tensor Definitions: Complete tensor mappings for MoE components including 128 routed experts + 1 shared expert
- Hybrid Layer Support: Added first_k_dense_replace parameter to handle different dense/MoE layer patterns between variants

## Model Loading (src/llama-model.cpp)

- Multi-variant Support: Automatic detection and loading for both 46-layer (Air) and 93-layer (full) models
- MoE Infrastructure: Complete expert weight loading with merged 3D tensor format
- Graph Implementation: New llm_build_glm45 class with sigmoid-based expert routing and top-8 selection
- Shared Expert Integration: Proper handling of shared expert computation alongside routed experts

## Conversion Support (convert_hf_to_gguf.py)

- HuggingFace Integration: Complete Glm45Model converter class with @ModelBase.register("Glm4MoeForCausalLM")
- Expert Tensor Merging: Sophisticated logic to merge expert weights into GGUF 3D tensor format
- Metadata Handling: Proper extraction and conversion of MoE parameters from HuggingFace config

## Technical Details

### MoE Architecture

- Expert Count: 128 routed experts + 1 shared expert per MoE layer
- Expert Selection: Top-8 experts per token with sigmoid-based routing (not softmax)
- Hybrid Layers: Dense layers for first N layers, MoE for remaining layers
- Weight Format: Expert weights stored as merged [num_experts, hidden_size, ffn_size] tensors

### Model Variants

- GLM-4.5: 355B total parameters, 32B active, 93 layers, includes K/Q norm tensors
- GLM-4.5-Air: 106B total parameters, 12B active, 46 layers, no K/Q norm tensors

### Testing

Builds successfully with no compilation errors.

Hopefully resolves #14921