@CISC 2025-07-29
> FYI when trying to run `convert_hf_to_gguf.py` on GLM4.5-Air-FP8, I get that some constants ending with `_EXPS` don't exist. If I replace these by `_EXP`, then I get a different error related to matrix mapping. Thank you for working on this!

That's because converting FP8 weights isn't supported yet, see #14810