@sammcj 2025-07-29
Note that preserving the nextn tensors does result in a larger GGUF (214GB -> 221GB for the f16)