@Thireus 2025-07-29
@sammcj, all tensors seem to be all there - https://huggingface.co/Thireus/GLM-4.5-THIREUS-BF16-SPECIAL_SPLIT/raw/main/tensors.map