@sammcj 2025-07-29
No worries at all, sorry if my message came across angry (it wasn't intended to), I just want to make sure that folks don't end up wasting their time or bandwidth, or worse - blame the llama.cpp project my lacking ability.

Thanks for wanting to help out, I _promise_ as soon as I think it's in a state that it could be tested I will update this thread.