@anikifoss 2025-07-29
No pressure. We're just all excited to try the new model. Thanks for doing the hard work!