@sammcj 2025-07-30
FYI @CISC, this is now working, but I don't want to waste your (or anyone else's) time - if the PR isn't up to scratch and will cost you more time than it'll save - feel free to reject.