@Noeda 2025-07-30
> FYI CISC, this is now working, but I don't want to waste your (or anyone else's) time - if the PR isn't up to scratch and will cost you more time than it'll save - feel free to reject.

@sammcj  The parts I looked through at least looks fine to me 😊 (I read the C++ parts in llama-model.cpp, I did not read the Python parts as of writing this comment, but might come back later to do that). In fact I find some of the comments better than what usually is on architectures e.g. (`// Check if this is a dense layer (n_layer_dense_lead=1, so layer 0 is dense)`), because it's a rare code comment that tells me what I should expect some variables to take as values as the code is running. Helps build mental model and check it's all good.

I went through code (the C++ parts) and did not find anything particularly objectionable other than some very inconsequential expert gating function parameter thing (and the non-issue of me getting confused at first about the layer). This doesn't tell me if the computation graph is correct or if some sub-optimal (in performance sense) code is used, but it looks good to me code quality-wise.