@Noeda 2025-07-30
> > @Noeda Did you test non-Air version yourself yet?
> 
> No sorry I haven't got the resources to try to run the big version (and slow Australian internet!).
> 
> I'll have a look over your gguf dump and logs and see if I can reverse engineer a fix.

Np np! If this PR is here next time I get to sit down (can't confirm when that'd be...but the model is interesting and I want it to work :-) ), I'll continue my troubleshooting. Meanwhile this can be a possible known issue, maybe someone else is able to also confirm it/see the issue.