@Noeda 2025-07-30
My final (hopefully useful) contribution of today, it *might* be these tensors that are present but not loaded:

```
blk.92.hnorm
blk.92.shared_head.head
blk.92.shared_head.norm
blk.92.enorm
blk.92.embed_tokens
blk.92.eh_proj
```

I had to rush to get that info out so I suggest double-checking that I didn't make a mistake...i.e. maybe last layer is handled incorrectly during loading.

<details>

<img width="1393" height="436" alt="Screenshot 2025-07-29 at 20 13 12" src="https://github.com/user-attachments/assets/be9aeea6-6de7-49b7-97f1-24b5c66c385e" />

</details>