@ddh0 2025-07-30
> tested the branch (pulled like an hour ago)
@csabakecskemeti there have been relevant changes since then, you will need to re-convert your model and see if the issue is fixed