@thad0ctor 2025-07-30
> > @thad0ctor I think the chat template needs to be flushed out that or my edits to quantize.cpp need some work because air is running wild at Q4
> 
> Yeah I just found this issue as well, some responses include tokens, e.g:
> 
> ```
> ## the fundamentals of graphic design thinking a principles of the basics design thinking|graphic design principles of the fundamental communication of organisation, the|>
> ```
> 
> This occurs even with the official jinja template, so I'll have to look into it further.
> 
> > @Noeda Did you test non-Air version yourself yet?
> 
> No sorry I haven't got the resources to try to run the big version (and slow Australian internet!).
> 
> I'll have a look over your gguf dump and logs and see if I can reverse engineer a fix.

good stuff, I can test the non-air version if needed too, just let me know what you need