@thad0ctor 2025-07-30
> re-testing with commit [07bb0dd](https://github.com/ggml-org/llama.cpp/commit/07bb0dd1e80f23aec522d774aaf29e2a3fcfc7c9) need few minutes

converting air to gguf with the commit as we speak, will report back