@csabakecskemeti 2025-07-30
f16 gguf with latest commit 
https://github.com/ggml-org/llama.cpp/commit/07bb0dd1e80f23aec522d774aaf29e2a3fcfc7c9
loads and generates but mainly repeates the prompt as before:

_how many 'r's are in the word strawberry? there are three 'r's in the word strawberry
First, the user asked: "how many 'r's are in the word strawberry?"

The user might have meant "how many 'r's are in the word strawberry?"

So, I need to count the number of 'r's in the word "strawberry"?

First, the user asked: "how many 'r's are in the word 'strawberry"?

The user might have meant "how many 'r's are in the word "strawberry"?

The user asked: "how many 'r'strawberry"?

The user's question is "how many 'r'strawberry"?_