@sammcj 2025-07-30
well, it's a lot better than it was, still has some issues with glitching as the conversation increases in length that need to be sorted.

```
$ git rev-parse HEAD
07bb0dd1e80f23aec522d774aaf29e2a3fcfc7c9

./bin/llama-server -m /Users/samm/LLM\ Models/zai-org_GLM-4.5-Air/glm-4.5-air-q3_K_M.gguf --temp 0.7 --top-k 20 --top-p 0.8 --min-p 0.0 --ctx-size 16384 --chat-template-file ./chat_template.jinja -ngl 99 -fa
```
<img width="1242" height="1087" alt="image" src="https://github.com/user-attachments/assets/7b318c3f-64bc-4046-b5fc-6ea2b425f14a" />