@csabakecskemeti 2025-07-30
GLM-4.5 
safetensor -> f16.gguf conversion completed

Generation:
```
...
model has unused tensor blk.92.eh_proj (size = 209715200 bytes) -- ignoring
model has unused tensor blk.92.embed_tokens (size = 3103784960 bytes) -- ignoring
model has unused tensor blk.92.enorm (size = 20480 bytes) -- ignoring
model has unused tensor blk.92.hnorm (size = 20480 bytes) -- ignoring
model has unused tensor blk.92.shared_head.head (size = 3103784960 bytes) -- ignoring
model has unused tensor blk.92.shared_head.norm (size = 20480 bytes) -- ignoring
load_tensors: tensor 'token_embd.weight' (f16) (and 1754 others) cannot be used with preferred buffer type CUDA_Host, using CPU instead
```
and it's just repeating the prompt

https://gist.github.com/csabakecskemeti/278ffb8ab8e4e13b605bfc1735bfed9b