@pwilkin 2025-07-30
Are you converting to bf16? The model's original format is bf16, not f16 - you might have artifacts if you're quantizing from f16.

Use --outtype bf16 on `convert_hf_to_gguf.py` to force bf16 conversion. I know people had problems with the big Ernie MoE model that were caused by quantizing from f16 instead of bf16.