@ubergarm 2025-08-02
@jukofyork 

> What perplexity does ik_llama.cpp give? We don't need to test the exact same file - just see if it gets such astronomically high values for comparison on any text.

Good question given I like to test both the `-fa` and non `-fa` path of a new model to make sure it "runs clean" with no nans and gives a reasonable perplexity to confirm a new implementation is not likely experiencing numerical issues.

I've successfully created an imatrix for both GLM-4.5 and the smaller GLM-4.5-Air BF16 GGUFs on ik_llama.cpp and uploaded to hf/ubergarm repos. Ran clean with no nans. That said, I'm not convinced things are working 100% there yet, and I haven't checked for updates regarding those MTP tensors which I ripped out temporarily just to get it going enough to test hah... There is a "small" test quant here which will only work on the ik_llama.cpp PR given a differencing in one of the tensor names: https://github.com/ikawrakow/ik_llama.cpp/pull/668/commits/d3d3fe626e6216f2d9d7cee629e14ea0146e8cc0#diff-ec77d8003b92ff283179456d36b8b56abf635e7b1232e70daf16676e8920ccf1R4099

Also just doing a quick perplexity test against wiki.test.raw. Seems like both the fa and non-fa paths are giving similar low perplexities over on ik_llama.cpp (unlike Hunyuan-A13B where the Pretrain had decent perplexity but the Instruct had 100x that).

<details>

<summary>👈 ik_llama.cpp llama-perplexity wik.test.raw</summary>

#### base command
```bash
#!/usr/bin/env bash

model=/mnt/raid/models/ubergarm/GLM-4.5-Air-GGUF/GLM-4.5-Air-Q8_0.gguf
# run with and without -fa and compare should have similar perplexities
numactl -N 1 -m 1 \
./build/bin/llama-perplexity \
    -m "$model" \
    -f wiki.test.raw \
    --seed 1337 \
    -fa \
    -fmoe \
    --ctx-size 512 \
    -ub 4096 -b 4096 \
    --numa numactl \
    --threads 128 \
    --threads-batch 192 \
    --no-mmap
```

#### With -fa
```bash
llama_new_context_with_model: attn_max_b = 0
llama_new_context_with_model: fused_moe  = 1
llama_new_context_with_model: ser        = -1, 0
llama_new_context_with_model: freq_base  = 1000000.0
llama_new_context_with_model: freq_scale = 1
llama_kv_cache_init:        CPU KV buffer size =   752.00 MiB
llama_new_context_with_model: KV self size  =  752.00 MiB, K (f16):  376.00 MiB, V (f16):  376.00 MiB
llama_new_context_with_model:        CPU  output buffer size =     4.63 MiB
llama_new_context_with_model:        CPU compute buffer size =  2432.00 MiB
llama_new_context_with_model: graph nodes  = 2196
llama_new_context_with_model: graph splits = 1

system_info: n_threads = 128 (n_threads_batch = 192) / 768 | AVX = 1 | AVX_VNNI = 1 | AVX2 = 1 | AVX512 = 1 | AVX512_VBMI = 1 | AVX512_VNNI = 1 | AVX512_BF16 = 1 | FMA = 1 | NEON = 0 | SVE = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 | MATMUL_INT8 = 0 | LLAMAFILE = 1 |
perplexity: tokenizing the input ..
perplexity: tokenization took 702.607 ms
perplexity: calculating perplexity over 565 chunks, n_ctx=512, batch_size=4096, n_seq=8
perplexity: 11.86 seconds per pass - ETA 13.95 minutes
[1]2.9683,[2]3.6421,[3]2.8255,[4]2.4795,[5]2.4274,[6]2.2993,[7]2.3185,[8]2.3825,[9]2.4789,[10]2.4829,[11]2.4803,[12]2.6836,[13]3.0113,[14]3.0670,[15]3.2206,[16]3.3005,[17]3.4273,[18]3.6616,[19]3.5959,[20]3.6366,[21]3.7236,[22]3.6939,[23]3.6089,[24]3.5393,[25]3.4925,[26]3.4686,[27]3.4417,[28]3.4798,[29]3.5207,[30]3.5761,[31]3.6409,[32]3.6938,[33]3.7615,[34]3.7993,[35]3.8870,[36]3.9409,[37]3.9626,[38]4.0556,[39]4.0904,[40]4.1254,[41]4.2084,[42]4.2248,[43]4.2271,[44]4.2660,[45]4.3763,[46]4.4387,[47]4.4054,[48]4.3349,[49]4.2864,[50]4.2921,[51]4.3365,[52]4.3709,[53]4.4215,[54]4.4532,[55]4.4749,[56]4.5071,[57]4.5010,[58]4.5212,[59]4.5320,[60]4.5802,[61]4.6283,[62]4.6877,[63]4.7401,[64]4.7672,[65]4.7822,[66]4.7940,[67]4.7964,[68]4.7994,[69]4.8427,[70]4.8448,[71]4.8309,[72]4.8314,[73]4.8455,[74]4.8726,[75]4.8733,[76]4.8040,[77]4.7385,[78]4.6907,[79]4.6435,[80]4.6086,[81]4.5644,[82]4.5478,[83]4.5457,[84]4.5224,[85]4.4977,[86]4.4737,[87]4.4695,[88]4.4515,[89]4.4145,[90]4.3925,[91]4.3613,[92]4.3489,[93]4.3107,[94]4.2989,[95]4.2717,[96]4.2772,[97]4.3091,[98]4.3000,[99]4.2736,[100]4.2513,[101]4.2634,[102]4.2737,[103]4.3053,[104]4.3360,[105]4.3703,[106]4.4083,[107]4.4834,[108]4.4926,[109]4.4962,[110]4.5388,[111]4.5606,[112]4.5232,[113]4.4975,[114]4.4958,[115]4.4810,[116]4.4774,[117]4.4662,[118]4.4607,[119]4.4465,[120]4.4318,[121]4.4172,[122]4.4093,[123]4.3743,[124]4.3589,[125]4.3528,[126]4.3354,[127]4.3247,[128]4.3305,[129]4.3364,[130]4.3408,[131]4.3512,[132]4.3548,[133]4.3550,[134]4.3592,[135]4.3777,[136]4.3861,[137]4.3975,[138]4.3976,[139]4.3946,[140]4.4109,[141]4.3998,[142]4.3989,[143]4.3905,[144]4.3789,[145]4.3656,[146]4.3584,[147]4.3558,[148]4.3605,[149]4.3546,[150]4.3561,[151]4.3406,[152]4.3313,[153]4.3259,[154]4.3161,[155]4.3106,[156]4.3060,[157]4.2997,[158]4.2933,[159]4.3017,[160]4.3085,[161]4.3125,[162]4.3265,[163]4.3247,[164]4.3381,[165]4.3630,[166]4.3929,[167]4.4213,[168]4.4358,[169]4.4717,[170]4.4967,[171]4.5181,[172]4.5499,[173]4.5386,[174]4.5135,[175]4.4944,[176]4.4707,[177]4.4499,[178]4.4324,[179]4.4102,[180]4.3915,[181]4.3843,[182]4.3981,[183]4.4193,[184]4.4409,
...
Final estimate: PPL = 4.6361 +/- 0.02821

llama_print_timings:        load time =   26491.36 ms
llama_print_timings:      sample time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =  792301.90 ms / 289280 tokens (    2.74 ms per token,   365.11 tokens per second)
llama_print_timings:        eval time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings:       total time =  810970.55 ms / 289281 tokens
```

#### Without -fa
```bash
llama_new_context_with_model: n_ctx      = 4096
llama_new_context_with_model: n_batch    = 4096
llama_new_context_with_model: n_ubatch   = 4096
llama_new_context_with_model: flash_attn = 0
llama_new_context_with_model: mla_attn   = 0
llama_new_context_with_model: attn_max_b = 0
llama_new_context_with_model: fused_moe  = 1
llama_new_context_with_model: ser        = -1, 0
llama_new_context_with_model: freq_base  = 1000000.0
llama_new_context_with_model: freq_scale = 1
llama_kv_cache_init:        CPU KV buffer size =   752.00 MiB
llama_new_context_with_model: KV self size  =  752.00 MiB, K (f16):  376.00 MiB, V (f16):  376.00 MiB
llama_new_context_with_model:        CPU  output buffer size =     4.63 MiB
llama_new_context_with_model:        CPU compute buffer size =  6784.05 MiB
llama_new_context_with_model: graph nodes  = 2383
llama_new_context_with_model: graph splits = 1

system_info: n_threads = 128 (n_threads_batch = 192) / 768 | AVX = 1 | AVX_VNNI = 1 | AVX2 = 1 | AVX512 = 1 | AVX512_VBMI = 1 | AVX512_VNNI = 1 | AVX512_BF16 = 1 | FMA = 1 | NEON = 0 | SVE = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 | MATMUL_INT8 = 0 | LLAMAFILE = 1 |
perplexity: tokenizing the input ..
perplexity: tokenization took 698.662 ms
perplexity: calculating perplexity over 565 chunks, n_ctx=512, batch_size=4096, n_seq=8
perplexity: 16.31 seconds per pass - ETA 19.20 minutes
[1]2.9722,[2]3.6290,[3]2.8172,[4]2.4727,[5]2.4342,[6]2.3065,[7]2.3281,[8]2.3811,[9]2.4785,[10]2.4798,[11]2.4724,[12]2.6734,[13]3.0026,[14]3.0594,[15]3.2139,[16]3.2936,[17]3.4211,[18]3.6551,[19]3.5883,[20]3.6295,[21]3.7176,[22]3.6877,[23]3.6052,[24]3.5371,[25]3.4898,[26]3.4628,[27]3.4359,[28]3.4741,[29]3.5176,[30]3.5738,[31]3.6382,[32]3.6908,[33]3.7593,[34]3.7969,[35]3.8848,[36]3.9393,[37]3.9608,[38]4.0527,[39]4.0869,[40]4.1221,[41]4.2044,[42]4.2202,[43]4.2231,[44]4.2618,[45]4.3718,[46]4.4349,[47]4.4018,[48]4.3325,[49]4.2834,[50]4.2891,[51]4.3335,[52]4.3678,[53]4.4196,[54]4.4510,[55]4.4725,[56]4.5051,[57]4.4995,[58]4.5198,[59]4.5300,[60]4.5783,[61]4.6261,[62]4.6856,[63]4.7386,[64]4.7658,[65]4.7808,[66]4.7932,[67]4.7957,[68]4.7983,[69]4.8409,[70]4.8435,[71]4.8295,[72]4.8290,[73]4.8427,[74]4.8697,[75]4.8701,[76]4.8014,[77]4.7358,[78]4.6879,[79]4.6409,[80]4.6061,[81]4.5614,[82]4.5452,[83]4.5436,[84]4.5212,[85]4.4962,[86]4.4707,[87]4.4670,[88]4.4485,[89]4.4108,[90]4.3892,[91]4.3576,[92]4.3472,[93]4.3095,[94]4.2972,[95]4.2701,[96]4.2762,[97]4.3084,[98]4.2988,[99]4.2732,[100]4.2505,[101]4.2625,[102]4.2725,[103]4.3040,[104]4.3341,[105]4.3682,[106]4.4062,[107]4.4818,[108]4.4910,[109]4.4945,[110]4.5369,[111]4.5586,[112]4.5211,[113]4.4956,[114]4.4941,[115]4.4792,[116]4.4759,[117]4.4645,[118]4.4590,[119]4.4448,[120]4.4301,[121]4.4152,[122]4.4074,[123]4.3722,[124]4.3569,[125]4.3511,[126]4.3337,[127]4.3235,[128]4.3293,[129]4.3353,[130]4.3397,[131]4.3501,[132]4.3534,[133]4.3531,[134]4.3575,[135]4.3759,[136]4.3840,[137]4.3952,[138]4.3952,[139]4.3921,[140]4.4083,[141]4.3974,[142]4.3965,[143]4.3879,[144]4.3764,[145]4.3631,[146]4.3566,[147]4.3537,[148]4.3584,[149]4.3529,[150]4.3545,[151]4.3390,[152]4.3300,[153]4.3244,[154]4.3147,[155]4.3094,[156]4.3047,[157]4.2986,[158]4.2921,[159]4.3004,[160]4.3072,[161]4.3113,[162]4.3261,[163]4.3244,[164]4.3378,[165]4.3628,[166]4.3928,[167]4.4210,[168]4.4355,
...
TODO WILL UPDATE IN  SEC
```

</details>